refactor(frontend): consolidate metadata handling into PageMetadata component
- Rename page-metadata.tsx to PageMetadata.tsx for consistency with export name
- Remove duplicate setPageMeta/resetPageMeta utilities from utils.ts
- Consolidate all document-head tag logic into single PageMetadata component
- Update imports across frontend to use new component path
- Remove unused DOM-based metadata mutation tests

Resolves #1002
